Move class definitions to a separate file, geom.py.
This is to make them accessible to translation functions.
